+++ src/raw_pixels.rs
@@ -1,6 +1,6 @@
 use embedded_graphics::prelude::*;
 
-use crate::{header::Bpp, raw_bmp::RawBmp};
+use crate::{header::Bpp, header::RowOrder, raw_bmp::RawBmp};
 
 /// Iterator over individual BMP pixels.
 ///
@@ -44,7 +44,7 @@ impl Iterator for RawPixels<'_, '_> {
         let p = self.position;
 
         if self.position.x == 0 {
-            let row_index = if self.raw_bmp.header().image_data_top_down {
+            let row_index = if self.raw_bmp.header().row_order == RowOrder::TopDown {
                 self.position.y
             } else {
                 (self.raw_bmp.size().height as i32 - 1) - self.position.y